+++ backend/app.py
@@ -1,10 +1,10 @@
-from flask import Flask, jsonify, request
-from flask_sqlalchemy import SQLAlchemy
+from flask import Flask
 from sqlalchemy import create_engine
-from sqlalchemy import text
-from sqlalchemy.exc import SQLAlchemyError
 import os
 from dotenv import load_dotenv
+from routes.rooms import create_rooms_router
+from routes.reserves import create_reserves_router
+from routes.users import create_users_router
 
 load_dotenv()
 
@@ -17,370 +17,9 @@ db =  os.getenv('DB_NAME')
 app = Flask(__name__)
 engine = create_engine(f'mysql+mysqlconnector://{user}:{password}@{host}/{db}')
 
-
-@app.route('/rooms', methods = ['GET'])
-def rooms():
-    conn = engine.connect()
-    
-    query = "SELECT * FROM rooms;"
-    try:
-        result = conn.execute(text(query))
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    
-
-    data = []
-    for row in result:
-        entity = {}
-        entity['id_room'] = row.id_room
-        entity['name'] = row.name
-        entity['capacity'] = row.capacity
-        entity['price'] = row.price
-        entity['stars'] = row.stars
-        data.append(entity)
-
-    return jsonify(data), 200
-
-
-@app.route('/rooms', methods = ['POST'])
-def create_room():
-    conn = engine.connect()
-    new_user = request.get_json()
-    query = f"""INSERT INTO rooms (name, capacity, price, stars) VALUES { new_user["name"] ,new_user["capacity"],new_user["price"],new_user["stars"]};"""
-    try:
-        conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify({'message': 'Se ha producido un error' + str(err.__cause__)})
-    
-    return jsonify({'message': 'se ha agregado la habitacion correctamente' }), 201
-
-
-
-@app.route('/rooms/<id>', methods = ['PATCH'])
-def update_room(id):
-    conn = engine.connect()
-    mod_user = request.get_json()
-
-    if not mod_user:
-        return ValueError("Ingrese los datos a actualizar")
-
-    set_pairs = []
-
-    for key, value in mod_user.items():
-        set_pairs.append(f"{key} = '{value}'")
-
-    set_clause = ", ".join(set_pairs)
-
-    query = f"UPDATE rooms SET {set_clause} WHERE id_room = {id}"
-
-
-    query_validation = f"SELECT * FROM rooms WHERE id_room = {id};"
-    try:
-        val_result = conn.execute(text(query_validation))
-        if val_result.rowcount!=0:
-            result = conn.execute(text(query))
-
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({'message': "La habitacion no existe"}), 404
-    except SQLAlchemyError as err:
-        return jsonify({'message': str(err.__cause__)})
-    return jsonify({'message': 'se ha modificado correctamente la habitacion '}), 200
-
-@app.route('/rooms/<id>', methods = ['GET'])
-def get_room(id):
-    conn = engine.connect()
-    query = f"""SELECT *
-            FROM rooms
-            WHERE id_room = {id};
-            """
-    try:
-        result = conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    if result.rowcount !=0:
-        data = {}
-        row = result.first()
-        data['id_room'] = row[0]
-        data['name'] = row[1]
-        data['capacity'] = row[2]
-        data['price'] = row[3]
-        data['stars'] = row[4]
-        return jsonify(data), 200
-    return jsonify({"message": f"La habitacion con id {id} no existe"}), 404
-
-
-@app.route('/rooms/<id>', methods = ['DELETE'])
-def delete_room(id):
-    conn = engine.connect()
-    query = f"""DELETE FROM rooms
-            WHERE id_room = {id};
-            """
-    validation_query = f"SELECT * FROM rooms WHERE id_room = {id}"
-    try:
-        val_result = conn.execute(text(validation_query))
-        if val_result.rowcount != 0 :
-            result = conn.execute(text(query))
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({"message": f"la habitacion con el id {id} no existe"}), 404
-    except SQLAlchemyError as err:
-        jsonify(str(err.__cause__))
-    return jsonify({'message': 'Se ha eliminado correctamente la habitacion'}), 202
-
-
-########--------------------------CRUD USERS---------------------------########
-
-@app.route('/users', methods = ['GET'])
-def users():
-    conn = engine.connect()
-    
-    query = "SELECT * FROM users;"
-    try:
-        result = conn.execute(text(query))
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    
-
-    data = []
-    for row in result:
-        entity = {}
-        entity['id_user'] = row.id_user
-        entity['email'] = row.email
-        entity['password'] = row.password
-        entity['name'] = row.name
-        entity['admin'] = row.admin
-        entity['created_at'] = row.created_at
-        data.append(entity)
-
-    return jsonify(data), 200
-
-
-@app.route('/users', methods = ['POST'])
-def create_user():
-    conn = engine.connect()
-    new_user = request.get_json()
-    query = f"""INSERT INTO users (email, password, name, admin) VALUES { new_user["email"] ,new_user["password"],new_user["name"],new_user["admin"]};"""
-    try:
-        conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify({'message': 'Se ha producido un error' + str(err.__cause__)})
-    
-    return jsonify({'message': 'se ha agregado el usuario correctamente' }), 201
-
-
-
-@app.route('/users/<id>', methods = ['PATCH'])
-def update_user(id):
-    conn = engine.connect()
-    mod_user = request.get_json()
-
-    if not mod_user:
-        return ValueError("Ingrese los datos a actualizar")
-
-    set_pairs = []
-
-    for key, value in mod_user.items():
-        set_pairs.append(f"{key} = '{value}'")
-
-    set_clause = ", ".join(set_pairs)
-
-    query = f"UPDATE users SET {set_clause} WHERE id_user = {id}"
-
-    query_validation = f"SELECT * FROM users WHERE id_user = {id};"
-    try:
-        val_result = conn.execute(text(query_validation))
-        if val_result.rowcount!=0:
-            result = conn.execute(text(query))
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({'message': "El usuario no existe"}), 404
-    except SQLAlchemyError as err:
-        return jsonify({'message': str(err.__cause__)})
-    return jsonify({'message': 'se ha modificado correctamente el usuario '}), 200
-
-@app.route('/users/<id>', methods = ['GET'])
-def get_user(id):
-    conn = engine.connect()
-    query = f"""SELECT *
-            FROM users
-            WHERE id_user = {id};
-            """
-    try:
-        result = conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    if result.rowcount !=0:
-        data = {}
-        row = result.first()
-        data['id_user'] = row[0]
-        data['name'] = row[1]
-        data['password'] = row[2]
-        data['email'] = row[3]
-        data['admin'] = row[4]
-        data['created_at'] = row[5]
-        return jsonify(data), 200
-    return jsonify({"message": "El usuario no existe"}), 404
-
-
-@app.route('/users/<id>', methods = ['DELETE'])
-def delete_user(id):
-    conn = engine.connect()
-    query = f"""DELETE FROM users
-            WHERE id_user = {id};
-            """
-    validation_query = f"SELECT * FROM users WHERE id_user = {id}"
-    try:
-        val_result = conn.execute(text(validation_query))
-        if val_result.rowcount != 0 :
-            result = conn.execute(text(query))
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({"message": "El usuario no existe"}), 404
-    except SQLAlchemyError as err:
-        jsonify(str(err.__cause__))
-    return jsonify({'message': 'Se ha eliminado correctamente al usuario'}), 202
-
-### CRUD reserves
-
-@app.route('/reserves',methods =['GET'])
-def reserves():
-    conn = engine.connect()
-
-    query = "SELECT * FROM reserves;"        
-    try:
-        result = conn.execute(text(query))
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    
-    
-    data = []
-    for row in result:
-        entity = {}
-        entity['id_reserve'] = row.id_reserve
-        entity['id_room'] = row.id_room
-        entity['id_user'] = row.id_user
-        entity['start_date']=row.start_date
-        entity['end_date'] = row.end_date
-        entity['created_at'] = row.created_at
-        entity['modified_at'] = row.modified_at
-        data.append(entity)
-
-    return jsonify(data), 200
-
-@app.route('/reserves', methods = ['POST'])
-def create_reserve():
-    conn = engine.connect()
-    new_reserve = request.get_json()
-    query = f"""INSERT INTO reserves (id_user,id_room,start_date, end_date) VALUES ('{new_reserve["id_user"]}', '{new_reserve["id_room"]}', '{new_reserve["start_date"]}', '{new_reserve["end_date"]}'); """
-    
-    try:
-        result = conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify({'message': 'Se ha producido un error' + str(err.__cause__)})
-    
-    return jsonify({'message': 'se ha creado la reserva correctamente' }), 201
-
-@app.route('/reserves/<id_reserve>', methods = ['PATCH'])
-def update_reserve(id_reserve):
-    conn = engine.connect()
-    mod_user = request.get_json()
-
-    if not mod_user:
-        return ValueError("Ingrese los datos a actualizar")
-
-    set_pairs = []
-
-    for key, value in mod_user.items():
-        set_pairs.append(f"{key} = '{value}'")
-
-    set_clause = ", ".join(set_pairs)
-
-    query = f"UPDATE reserves SET {set_clause} WHERE id_reserve = {id_reserve}"
-
-    query_validation = f"SELECT * FROM reserves WHERE id_reserve = {id_reserve};"
-    try:
-        val_result = conn.execute(text(query_validation))
-        if val_result.rowcount!=0:
-            result = conn.execute(text(query))
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({'message': "El numero de reserva no existe"}), 404
-    except SQLAlchemyError as err:
-        return jsonify({'message': str(err.__cause__)})
-    return jsonify({'message': 'Se ha modificado la reserva correctamente'}), 200
-
-@app.route('/reserves/<id_reserve>', methods = ['GET'])
-def get_reserve(id_reserve):
-    conn = engine.connect()
-    query = f"""SELECT *
-            FROM reserves
-            WHERE id_reserve = {id_reserve};
-            """
-    try:
-        result = conn.execute(text(query))
-        conn.commit()
-        conn.close()
-    except SQLAlchemyError as err:
-        return jsonify(str(err.__cause__))
-    if result.rowcount !=0:
-        data = {}
-        row = result.first()
-        data['id_reserve'] = row[0]
-        data['id_user'] = row[1]
-        data['id_room'] = row[2]
-        data['start_date'] = row[3]
-        data['end_date'] = row[4]
-        data['created_at'] = row[5]
-        data['modified_at'] = row[6]
-        return jsonify(data), 200
-    return jsonify({"message": "La reserva no existe"}), 404
-
-
-@app.route('/reserves/<id_reserve>', methods = ['DELETE'])
-def delete_reserve(id_reserve):
-    conn = engine.connect()
-    query = f"""DELETE FROM reserves
-            WHERE id_reserve = {id_reserve};
-            """
-    validation_query = f"SELECT * FROM reserves WHERE id_reserve = {id_reserve}"
-    try:
-        val_result = conn.execute(text(validation_query))
-        if val_result.rowcount != 0 :
-            result = conn.execute(text(query))
-            conn.commit()
-            conn.close()
-        else:
-            conn.close()
-            return jsonify({"message": "La reserva no existe."}), 404
-    except SQLAlchemyError as err:
-        jsonify(str(err.__cause__))
-    return jsonify({'message': 'Se ha eliminado correctamente'}), 202
-
-
+app.register_blueprint(create_rooms_router(engine))
+app.register_blueprint(create_reserves_router(engine))
+app.register_blueprint(create_users_router(engine))
 
 if __name__ == "__main__":
     app.run("127.0.0.1", port="5000", debug=True)
